Refactor utils into directory
