+++ .eslintrc
@@ -1,0 +1,8 @@
+{
+  "extends": [
+    "standard"
+  ],
+  "rules": {
+    "no-var": "off"
+  }
+}
