+++ lib/constants.js
@@ -1,9 +1,18 @@
 'use strict'
 
+/**
+ * A set of property names that indicate the value represents an error object.
+ *
+ * @typedef {string[]} K_ERROR_LIKE_KEYS
+ */
+
 module.exports = {
   DATE_FORMAT: 'yyyy-mm-dd HH:MM:ss.l o',
   DATE_FORMAT_SIMPLE: 'HH:MM:ss.l',
 
+  /**
+   * @type {K_ERROR_LIKE_KEYS}
+   */
   ERROR_LIKE_KEYS: ['err', 'error'],
 
   MESSAGE_KEY: 'msg',
