+++ lib/utils/is-object.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = isObject
+
+function isObject (input) {
+  return Object.prototype.toString.apply(input) === '[object Object]'
+}
