+++ lib/utils/is-valid-date.js
@@ -1,0 +1,14 @@
+'use strict'
+
+module.exports = isValidDate
+
+/**
+ * Checks if the argument is a JS Date and not 'Invalid Date'.
+ *
+ * @param {Date} date The date to check.
+ *
+ * @returns {boolean} true if the argument is a JS Date and not 'Invalid Date'.
+ */
+function isValidDate (date) {
+  return date instanceof Date && !Number.isNaN(date.getTime())
+}
