+++ lib/utils/noop.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = function noop () {}
